requirements.txt
rope
pylint
autopep8
colorama
requests
psd-tools
python-dotenv